Add ICrudService and refactoring

+++ Theater.Core/ServiceBase.cs
@@ -10,16 +10,21 @@ using Theater.Entities;
 namespace Theater.Core
 {
     public abstract class ServiceBase<TModel, TEntity> : ICrudService<TModel, TEntity>
-        where TEntity : class, IEntity
+        where TEntity : class, IEntity, new()
         where TModel : class
     {
         protected readonly IMapper Mapper;
         protected readonly ICrudRepository<TEntity> Repository;
+        protected readonly IDocumentValidator<TModel> DocumentValidator;
 
-        protected ServiceBase(IMapper mapper, ICrudRepository<TEntity> repository)
+        protected ServiceBase(
+            IMapper mapper, 
+            ICrudRepository<TEntity> repository,
+            IDocumentValidator<TModel> documentValidator)
         {
             Mapper = mapper;
             Repository = repository;
+            DocumentValidator = documentValidator;
         }
 
         public async Task<WriteResult<TModel>> GetById(Guid id)
@@ -31,29 +36,58 @@ namespace Theater.Core
                 : WriteResult.FromValue(Mapper.Map<TModel>(entity));
         }
 
-        public async Task<WriteResult<DocumentMeta>> CreateOrUpdate(TModel model, Guid? pieceId)
+        public async Task<WriteResult<DocumentMeta>> CreateOrUpdate(TModel model, Guid? entityId)
         {
-            var piece = await Repository.GetByEntityId(pieceId ?? Guid.NewGuid());
+            var entity = entityId.HasValue 
+                ? await Repository.GetByEntityId(entityId.Value) 
+                : null;
 
-            if (piece is null)
-            {
-                piece = Mapper.Map<TEntity>(model);
-
-                await Repository.Add(piece);
-            }
-
-            Mapper.Map(model, piece);
-
-            await Repository.Update(piece);
-
-            return WriteResult.FromValue(new DocumentMeta(piece.Id));
+            return entity is null 
+                ? await CreateEntity(model) 
+                : await UpdateEntity(entity, model, entityId);
         }
 
         public async Task<WriteResult> Delete(Guid id)
         {
+            var validationResult = await DocumentValidator.CheckIfCanDelete(id);
+
+            if (!validationResult.IsSuccess)
+                return validationResult;
+
             await Repository.Delete(id);
 
             return WriteResult.Successful;
+        }
+
+        private async Task<WriteResult<DocumentMeta>> CreateEntity(TModel model)
+        {
+            var validationResult = await DocumentValidator.CheckIfCanCreate(model);
+
+            if (!validationResult.IsSuccess)
+                return WriteResult<DocumentMeta>.FromError(validationResult.Error);
+
+            var entity = Mapper.Map<TEntity>(model);
+
+            await Repository.Add(entity);
+
+            return WriteResult.FromValue(new DocumentMeta(entity.Id));
+        }
+
+        private async Task<WriteResult<DocumentMeta>> UpdateEntity(TEntity entity, TModel model, Guid? entityId)
+        {
+            if (!entityId.HasValue)
+                return WriteResult<DocumentMeta>.FromError(ErrorModel.Default("update-conflict", "При обновлении сущности необходимо передавать идентификатор "));
+
+            var validationResult = await DocumentValidator.CheckIfCanUpdate(entityId.Value, model);
+
+            if (validationResult.IsSuccess)
+                return WriteResult<DocumentMeta>.FromError(validationResult.Error);
+
+            Mapper.Map(model, entity);
+
+            await Repository.Update(entity);
+
+            return WriteResult.FromValue(new DocumentMeta(entity.Id));
         }
     }
 }
